Modernize printf() format
Use macro constants from <inttypes.h> to print width-based integral
types from <stdint.h>.

+++ src/occtl/unix.c
@@ -21,6 +21,7 @@
  */
 
 #include <config.h>
+#include <inttypes.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -127,8 +128,9 @@ static int send_cmd(struct unix_ctx *ctx, unsigned int cmd, const void *data,
 
 		if (msg_map[cmd] != rep->cmd) {
 			fprintf(stderr,
-				"Unexpected message '%d', expected '%d'\n",
-				(int)rep->cmd, (int)msg_map[cmd]);
+				"Unexpected message '%u', expected '%" PRIu8
+				"'\n",
+				rep->cmd, msg_map[cmd]);
 			ret = -1;
 			goto fail;
 		}
@@ -1759,8 +1761,9 @@ int handle_events_cmd(struct unix_ctx *ctx, const char *arg,
 
 		if (header[0] != CTL_CMD_TOP_UPDATE_REP) {
 			fprintf(stderr,
-				"events: Unexpected message '%d', expected '%d'\n",
-				(int)header[0], (int)CTL_CMD_TOP_UPDATE_REP);
+				"events: Unexpected message '%d', expected '%" PRIu8
+				"'\n",
+				header[0], CTL_CMD_TOP_UPDATE_REP);
 			ret = -1;
 			break;
 		}
@@ -1793,7 +1796,8 @@ int handle_events_cmd(struct unix_ctx *ctx, const char *arg,
 			common_info_cmd(rep2->user, stdout, params);
 		} else {
 			if (rep2->connected) {
-				printf("%s: connected user '%s' (%u) from %s with IP %s\n",
+				printf("%s: connected user '%s' (%" PRId32
+				       ") from %s with IP %s\n",
 				       rep2->user->user[0]->vhost,
 				       rep2->user->user[0]->username,
 				       rep2->user->user[0]->id,
@@ -1809,7 +1813,8 @@ int handle_events_cmd(struct unix_ctx *ctx, const char *arg,
 				print_time_ival7(
 					tmpbuf, time(NULL),
 					rep2->user->user[0]->conn_time);
-				printf("%s: disconnect user '%s' (%u) from %s with IP %s (reason: %s, time: %s)\n",
+				printf("%s: disconnect user '%s' (%" PRId32
+				       ") from %s with IP %s (reason: %s, time: %s)\n",
 				       rep2->user->user[0]->vhost,
 				       rep2->user->user[0]->username,
 				       rep2->user->user[0]->id,
