+++ src/worker-kkdcp.c
@@ -19,6 +19,7 @@
 
 #include <config.h>
 
+#include <inttypes.h>
 #include <stdlib.h>
 #include <stdarg.h>
 #include <stdio.h>
@@ -152,7 +153,7 @@ int post_kkdcp_handler(worker_st *ws, unsigned int http_ver)
 	ws_add_score_to_ip(ws, WSCONFIG(ws)->ban_points_kkdcp, 0, 0);
 	oclog(ws, LOG_HTTP_DEBUG,
 	      "HTTP processing kkdcp framed request: %u bytes",
-	      (unsigned int)req->body_length);
+	      req->body_length);
 
 	length = BUF_SIZE;
 	buf = talloc_size(ws, length);
@@ -199,7 +200,7 @@ int post_kkdcp_handler(worker_st *ws, unsigned int http_ver)
 	}
 
 	oclog(ws, LOG_HTTP_DEBUG, "HTTP sending kkdcp request: %u bytes",
-	      (unsigned int)length);
+	      length);
 	ret = send(fd, buf, length, 0);
 	if (ret != length) {
 		if (ret == -1) {
@@ -238,8 +239,9 @@ int post_kkdcp_handler(worker_st *ws, unsigned int http_ver)
 		memcpy(&mlength, buf, 4);
 		mlength = ntohl(mlength);
 		if (mlength >= BUF_SIZE - 4) {
-			oclog(ws, LOG_ERR, "kkdcp: too long message (%d bytes)",
-			      (int)mlength);
+			oclog(ws, LOG_ERR,
+			      "kkdcp: too long message (%" PRIu32 " bytes)",
+			      mlength);
 			reason = "kkdcp: error receiving from server";
 			goto fail;
 		}
@@ -257,7 +259,7 @@ int post_kkdcp_handler(worker_st *ws, unsigned int http_ver)
 	}
 
 	oclog(ws, LOG_HTTP_DEBUG, "HTTP processing kkdcp reply: %u bytes",
-	      (unsigned int)length);
+	      length);
 
 	cstp_cork(ws);
 	ret = cstp_printf(ws, "HTTP/1.%u 200 OK\r\n", http_ver);
@@ -279,8 +281,8 @@ int post_kkdcp_handler(worker_st *ws, unsigned int http_ver)
 	}
 
 	oclog(ws, LOG_HTTP_DEBUG, "HTTP sending kkdcp framed reply: %u bytes",
-	      (unsigned int)length);
-	ret = cstp_printf(ws, "Content-Length: %u\r\n", (unsigned int)length);
+	      length);
+	ret = cstp_printf(ws, "Content-Length: %u\r\n", length);
 	if (ret < 0) {
 		goto fail;
 	}
